feat: new passkey signature format (#58)
* feat: new passkey signature format

includes the credential id, which is hopefully available

* fix: update to include credential id on creation

Also update contracts locally

* chore: update to latest contracts

* fix: update lockfile from contract dep change

* fix: import for utils

* fix: update signature format tests

It doesn't look correct visually, but it's mocked on both ends
so it's not super clear on what it's testing

* feat: attempt to use new dynamic contract addresses

When deploying the contracts, it should create the file, blocking the build

* fix: provide an example file

Can have a different test to make sure this gets updated or
put the contract build step into the other builds

* fix: add create2 to nft deploy

Try encoding the local contracts directly instead of via env vars
doesn't appear to work locally...

* fix: encode passkey id as bytes

This was correct on creation, but wrong on signing

* chore: update to latest contract

* fix: update hardcoded signatures for passkey view

This was super tough to find and would have been a compile error
had we imported the abi directly from the contracts!

This is something I'm going to look into so we don't lose our minds

* fix: remove debugging changes

You just have to know that the quest address is added via .env.local
so the object in the config isn't used

* feat: attempt to import abi automatically

Having to manually update the ABI kinda defeats the point of type
safety, but it's clear the auto-generated one doesn't play nice
with the existing file so more work will be necessary to update it
so it's compatible

* fix: use view function to get key

Picks up a fix from the publish function for the correct abi

* fix: remove publish from CI

It looks like nuxt builds don't like the poorly formatted json,
so instead of trying to pretty print the abi automatically we'll
rely on the editor tools to do this.

At the very least the publish step is documented?

* fix: update import names

problem was not the file format, but the file name change

* fix: use pnpm run build instead of build

CI is behaving strange?

* fix: use correct contract branch

* fix: pull contract directory in CI

This explains why the build was failing

* feat: setup local node for deploy

This should be fine because we aren't actually using it

* [wip] Add application metadata (#82)

* feat: provide auth-server metadata

Can be helpful for custom auth-server implementions that want extra data

* fix: use a record instead of an array of records

There's not really much extra flexibility here, so just keep it simple

* fix: update demo app with new metadata format

* Update packages/auth-server/components/session/Metadata.vue

---------

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>

* fix: update to breaking factory change

This was already fixed during guardian module work

* fix: update ABI

breaking name and argument changes were not published
Added publish step to readme

* fix: update new variable name

now automatically checked via types!

* fix: update test mock

The log parsing format is broken because the return type changed,
but regenerating the log format is non-trival so that can be updated later.

introducing #84

* fix: update account id length

Updating the tests to match

---------

Co-authored-by: Nicolas Villanueva <1890113+MexicanAce@users.noreply.github.com>