ECDSA client examples (#115)
* feat: ECDSA client with single owner

* chore: remove console.log

* feat: update aa factory abi

* fix: import types

* feat: add examples as ecdsa docs

So this fails when attempting to send with the error message:
"eth_signTypedData_v4" does not exist / is not available.

But the rest of it looks great in terms of usage examples!

* fix: linting and import

Not sure how it fails the sdk, since it works locally?

* chore: add another signer type

It's not clear this is helpful, but I'm quite stumped on how this keeps failing
in CI

* fix: use account instead of ecdsa wallet signer

This fixes the missing method on the signer

* fix: code review comments

---------

Co-authored-by: Sulpiride <sobirovutkir@gmail.com>