feat: update readme with new publish steps (#129)
* feat: update readme with new publish steps

Adding the automatic formatting post-publish reduces git diff noise
The name change is because the old file was manually named and imported

* fix: update pnpm lockfile

* fix: update filename

The import was renamed, but the filename was missed!

* chore: add run command to build

pnpm I think has the npm compat built in where the special scripts
can run without the run command.

It's probably better to not rely on that!