fix: test order updated
Now mocking logs since the order of test execution could leave the mocks
in a bad state.
This is an awful, awful problem to have with mocks but it's not clear
that resetting manages the state correctly, so only overwriting it
provides a sane end-state behavior.

These tests are probably a net loss in terms of being more mock than
test.