fix: deploy new factory with old sso beacon
The problem was in the upgrade process a new SSO beacon proxy was
deployed for the new implemention instead of re-using the existing one.
This was because it used the era-sepolia.json which only contains the
contracts necessary for the auth-server and didn't have the SSO Beacon
address.
The SSO beacon address needs to stay the same because its address is
specifically allowed by the sequencer.

Might also need to update + fund the paymaster, will test in PR.