chore: failing tests
back to where I was before, but now with more steps and a worse UI.
Might even try to import the old UI and ask it to update for viem
clients.
Or might just try to rebase and start on the tests fresh, at least we're
not getting stuck on the wasm build issues yet, it looks pretty simple
so far!